% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepNeuralNetwork.predict.r
\name{deepNeuralNetwork.predict}
\alias{deepNeuralNetwork.predict}
\title{Calculate a prediction from a trained DeepNNModel.}
\usage{
deepNeuralNetwork.predict(model.trained, data = X.test,
  standarization = NULL)
}
\arguments{
\item{model.trained}{An object of class DeepNNModel containing a trained model.}

\item{data}{Matrix containing the test data without the observed variable that we want to predict.}

\item{standarization}{Character or list. Standarization method to be used. If a list of features (corresponding to rows in the input matrix) is
supplied, a standarization feature-z-score is done. "r"=robust median z-score. "s"=standar mean z-score.}
}
\value{
A vector or matrix containing the predicted values.
}
\description{
`deepNeuralNetwork.predict()` returns a prediction over the given data.
}
\details{
This function will calculate a prediction using the supplied data. It performs the selected data standarization
(beware that the input data must be standarized using the same method as in training) and returns the predicted values.
}
