% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepNeuralNetwork.run.r
\name{deepNeuralNetwork.run}
\alias{deepNeuralNetwork.run}
\title{Partial Prediction for in-training use.}
\usage{
deepNeuralNetwork.run(model.trained, data = X.test)
}
\arguments{
\item{model.trained}{An object of class DeepNNModel.}

\item{data}{Matrix containing the test data without the observed variable that we want to predict.}
}
\value{
A vector or matrix containing the prediction values.
}
\description{
`deepNeuralNetwork.run()` returns a partial prediction over the test data to measure overfitting if test data is supplied by the user.
It is calculated using the actual in-training-model.
}
\details{
This function will calculate a prediction using test data (if it is supplied by the user). [deepNeuralNetwork.training] function will use this function
and depending on the result given it can decide whether to continue with the training or stop it.
}
