% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepNeuralNetwork.save.r
\name{deepNeuralNetwork.save}
\alias{deepNeuralNetwork.save}
\title{Saves a DeepNNModel object to a file.}
\usage{
deepNeuralNetwork.save(wn = NULL, nameoffile = paste("DNN.model.",
  format(Sys.time(), "\%d.\%m.\%Y.\%H.\%M.\%S"), sep = ""))
}
\arguments{
\item{wn}{Object of class DeepNNModel.}

\item{nameoffile}{Character. The name of the file.}
}
\value{
If success the name of the file.
}
\description{
`deepNeuralNetwork.save()` Saves a DeepNNModel object to a .RData file.
}
\details{
`deepNeuralNetwork.save()` Saves a DeepNNModel object to a .RData file.
}
