% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepNeuralNetwork.standarize.r
\name{deepNeuralNetwork.standarize}
\alias{deepNeuralNetwork.standarize}
\title{Standarizes the input data.}
\usage{
deepNeuralNetwork.standarize(x = NULL, method = "r")
}
\arguments{
\item{x}{Matrix. The data to be standarized.}

\item{method}{Character. Specifies the method used. "s"=standar z-score using the mean. "r"=robust z-score using the median.}
}
\value{
The standarized data matrix.
}
\description{
`deepNeuralNetwork.standarize()` standarizes the given data by calculating either a z-score using the mean and the sd or
a robust z-score using the median and the MAD.
}
\details{
`deepNeuralNetwork.standarize()` standarizes the given data by calculating either a z-score using the mean and the sd or
a robust z-score using the median and the MAD.
}
