% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepNeuralNetwork.standarizegenescore.r
\name{deepNeuralNetwork.standarizegenescore}
\alias{deepNeuralNetwork.standarizegenescore}
\title{Standarizes the input data using a feature as a target.}
\usage{
deepNeuralNetwork.standarizegenescore(x = NULL, method = "gene",
  gene.list = NA)
}
\arguments{
\item{x}{Matrix. The data to be standarized.}

\item{method}{Character. Only gene is acepted.}

\item{gene.list}{Character vector. List of features to use if method is "genecomb".}
}
\value{
The standarized data matrix.
}
\description{
`deepNeuralNetwork.standarizegenescore()` standarizes the given data by calculating a personalized z-score using the feature as centroid and dividing by the MAD.
}
\details{
`deepNeuralNetwork.standarizegenescore()` standarizes the given data by calculating a personalized z-score using the feature as centroid and dividing by the MAD.
}
