% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDeepNN.r
\name{draw.dnn}
\alias{draw.dnn}
\title{Plot of the DNN}
\usage{
draw.dnn(NperL = 0, biasVector)
}
\arguments{
\item{NperL}{Vector. The layers and the number of neurons in each.}

\item{biasVector}{NA. Deprecated.}
}
\description{
This function plots the DNN arquitecture.
}
