% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearPlot.r
\name{mplot_lineal}
\alias{mplot_lineal}
\title{Linear Regression Results Plot}
\usage{
mplot_lineal(observed, predicted, title = "Regression Model Results",
  subtitle = NA, model_name = NA, save = FALSE, subdir = NA,
  deviation = FALSE, transparent = FALSE, x.lim = c(0, 100),
  y.lim = c(0, 100), x.lab = "Observed values (Y)",
  y.lab = "Predicted values (Y')", file_name = "viz_lineal.png")
}
\arguments{
\item{observed}{Vector. Observed values: real data values label.}

\item{predicted}{Vector. Predicted value or model's result.}

\item{title}{Character. Title to be shown in the plot.}

\item{subtitle}{Character. Subtitle to be shown in the plot.}

\item{model_name}{Character. Model's name}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{deviation}{Boolean. Plot the deviation?}

\item{transparent}{Numeric. Transparency parameter.}

\item{x.lim}{Vector. The bottom and upper limit of the values on X axis.}

\item{y.lim}{Vector. The bottom and upper limit of the values on Y axis.}

\item{x.lab}{Character. The label to be printed on the X axis plot.}

\item{y.lab}{Character. The label to be printed on the Y axis plot.}

\item{file_name}{Character. File name as you wish to save the plot}
}
\description{
This function plots a Linear Regression Result
}
