% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.plot.r
\name{residuals.plot}
\alias{residuals.plot}
\title{Plot of the residuals Results}
\usage{
\method{residuals}{plot}(observed = NULL, predicted = NULL,
  title = "Residuals plot")
}
\arguments{
\item{observed}{Vector. The observed values from the data training.}

\item{predicted}{Vector. The predicted values from the model's result.}

\item{title}{Character. The title to be shown in the plot.}
}
\description{
This function plots the residuals from the results once the training has finish.
}
