% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.data.r
\docType{data}
\name{training.data}
\alias{training.data}
\title{Gene expression of 3 different human brain datasets of cortex region for training a Deep Neural Network.}
\format{A data.frame of class "data.frame"}
\source{
Datasets are available publicy at GEO: GSE25219, GSE46706, GSE48350
}
\usage{
data(training.data)
}
\description{
Data from 3 different datasets of human cortex. Gene expression of 1049 genes selected
because its implication in human aging (Oscar González-Velasco, et al., BBA - Gene Regulatory Mechanisms, https://doi.org/10.1016/j.bbagrm.2020.194491).
All 3 datasets are from Affymetrix HG 2.0 microarrays.
Gene expression signal is normalized using RMA and it includes the age of every individual.
}
\examples{
data(training.data)
head(training.data)

}
\references{
Datasets are available publicy at GEO: GSE25219, GSE46706, GSE48350
}
\keyword{datasets}
